refactor(rpc): isolate stream notifications
