+++ services/local-service/internal/rpc/server.go
@@ -3,13 +3,8 @@ package rpc
 
 import (
 	"context"
-	"encoding/json"
 	"errors"
-	"io"
-	"net"
 	"net/http"
-	"strings"
-	"sync"
 	"time"
 
 	serviceconfig "github.com/cialloclaw/cialloclaw/services/local-service/internal/config"
@@ -101,126 +96,6 @@ func (s *Server) Shutdown(ctx context.Context) error {
 	return nil
 }
 
-// handleStreamConn serves JSON-RPC requests on a long-lived stream and then
-// replays buffered notifications on the same connection.
-func (s *Server) handleStreamConn(conn net.Conn) {
-	defer conn.Close()
-
-	decoder := json.NewDecoder(conn)
-	encoder := json.NewEncoder(conn)
-	var writeMu sync.Mutex
-
-	for {
-		var request requestEnvelope
-		if err := decoder.Decode(&request); err != nil {
-			if errors.Is(err, io.EOF) {
-				return
-			}
-
-			_ = encoder.Encode(newErrorEnvelope(nil, &rpcError{
-				Code:    errInvalidParams,
-				Message: "INVALID_PARAMS",
-				Detail:  "invalid json-rpc payload",
-				TraceID: "trace_rpc_decode",
-			}))
-			return
-		}
-
-		streamedRuntimeCounts := map[string]int{}
-		requestTaskIDs, requestSessionID, requestTraceID := requestRoutingHints(request)
-		var requestTaskMu sync.RWMutex
-		addRequestTaskID := func(taskID string) {
-			trimmed := strings.TrimSpace(taskID)
-			if trimmed == "" {
-				return
-			}
-			requestTaskMu.Lock()
-			if requestTaskIDs == nil {
-				requestTaskIDs = map[string]bool{}
-			}
-			requestTaskIDs[trimmed] = true
-			requestTaskMu.Unlock()
-		}
-		hasRequestTaskID := func(taskID string) bool {
-			requestTaskMu.RLock()
-			defer requestTaskMu.RUnlock()
-			return requestTaskIDs != nil && requestTaskIDs[taskID]
-		}
-		matchesTaskStart := func(sessionID, traceID string) bool {
-			switch {
-			case requestTraceID != "":
-				return requestTraceID == traceID
-			case requestSessionID != "":
-				return requestSessionID == sessionID
-			default:
-				return false
-			}
-		}
-
-		unsubscribeRuntime := func() {}
-		if requestTaskIDs != nil || shouldTrackStartedTask(request.Method) {
-			unsubscribeRuntime = s.orchestrator.SubscribeRuntimeNotifications(func(taskID string, method string, params map[string]any) {
-				if !isLiveRuntimeMethod(method) {
-					return
-				}
-				notificationTaskID := runtimeNotificationTaskID(taskID, params)
-				if notificationTaskID == "" || !hasRequestTaskID(notificationTaskID) {
-					return
-				}
-				writeMu.Lock()
-				defer writeMu.Unlock()
-				if err := encoder.Encode(newNotificationEnvelope(method, params)); err == nil {
-					streamedRuntimeCounts[notificationKey(method, notificationTaskID, params)]++
-				}
-			})
-		}
-
-		unsubscribeTaskStart := func() {}
-		if shouldTrackStartedTask(request.Method) {
-			unsubscribeTaskStart = s.orchestrator.SubscribeTaskStarts(func(taskID, sessionID, traceID string) {
-				if !matchesTaskStart(sessionID, traceID) {
-					return
-				}
-				addRequestTaskID(taskID)
-			})
-		}
-
-		response := s.dispatch(request)
-		unsubscribeTaskStart()
-		unsubscribeRuntime()
-
-		writeMu.Lock()
-		err := encoder.Encode(response)
-		writeMu.Unlock()
-		if err != nil {
-			return
-		}
-
-		for _, taskID := range taskIDsFromResponse(response) {
-			notifications, err := s.orchestrator.DrainNotifications(taskID)
-			if err != nil {
-				continue
-			}
-
-			for _, notification := range notifications {
-				method := stringValue(notification, "method", "task.updated")
-				params := mapValue(notification, "params")
-				key := notificationKey(method, taskID, params)
-				if isLiveRuntimeMethod(method) && streamedRuntimeCounts[key] > 0 {
-					streamedRuntimeCounts[key]--
-					continue
-				}
-				writeMu.Lock()
-				err := encoder.Encode(newNotificationEnvelope(method, params))
-				writeMu.Unlock()
-				if err != nil {
-					return
-				}
-			}
-		}
-	}
-}
-
 // dispatch is the single RPC dispatch path that validates protocol shape,
 // resolves handlers, decodes params, and rewraps orchestrator output.
 func (s *Server) dispatch(request requestEnvelope) any {
@@ -259,124 +134,4 @@ func (s *Server) dispatch(request requestEnvelope) any {
 // nowRFC3339 returns the unified response timestamp format.
 func (s *Server) nowRFC3339() string {
 	return s.now().Format(time.RFC3339)
-}
-
-// taskIDsFromResponse recursively collects task_id values from a success
-// response so the transport can replay related notifications afterward.
-func taskIDsFromResponse(response any) []string {
-	success, ok := response.(successEnvelope)
-	if !ok {
-		return nil
-	}
-
-	ids := map[string]struct{}{}
-	collectTaskIDs(success.Result.Data, ids)
-
-	result := make([]string, 0, len(ids))
-	for taskID := range ids {
-		result = append(result, taskID)
-	}
-
-	return result
-}
-
-func requestRoutingHints(request requestEnvelope) (map[string]bool, string, string) {
-	params, rpcErr := decodeParams(request.Params)
-	if rpcErr != nil {
-		return nil, "", ""
-	}
-
-	ids := map[string]struct{}{}
-	collectTaskIDs(params, ids)
-	var result map[string]bool
-	if len(ids) > 0 {
-		result = make(map[string]bool, len(ids))
-		for taskID := range ids {
-			result[taskID] = true
-		}
-	}
-	return result, stringValue(params, "session_id", ""), stringValue(mapValue(params, "request_meta"), "trace_id", "")
-}
-
-func shouldTrackStartedTask(method string) bool {
-	return method == "agent.task.start" || method == "agent.input.submit"
-}
-
-func isLiveRuntimeMethod(method string) bool {
-	return strings.HasPrefix(method, "loop.") || method == "task.steered"
-}
-
-func runtimeNotificationTaskID(taskID string, params map[string]any) string {
-	if strings.TrimSpace(taskID) != "" {
-		return taskID
-	}
-	if params == nil {
-		return ""
-	}
-	rawTaskID, _ := params["task_id"].(string)
-	return strings.TrimSpace(rawTaskID)
-}
-
-func notificationKey(method, taskID string, params map[string]any) string {
-	encoded, err := json.Marshal(normalizeNotificationKey(method, taskID, params))
-	if err != nil {
-		return method
-	}
-	return method + ":" + string(encoded)
-}
-
-func normalizeNotificationKey(method, taskID string, params map[string]any) map[string]any {
-	if !isLiveRuntimeMethod(method) {
-		return map[string]any{
-			"task_id": strings.TrimSpace(taskID),
-			"params":  params,
-		}
-	}
-
-	normalizedTaskID := strings.TrimSpace(taskID)
-	if normalizedTaskID == "" {
-		normalizedTaskID = runtimeNotificationTaskID("", params)
-	}
-
-	payload := map[string]any{}
-	if event := mapValue(params, "event"); len(event) > 0 {
-		payload = mapValue(event, "payload")
-	} else {
-		for key, value := range params {
-			if key == "task_id" {
-				continue
-			}
-			payload[key] = value
-		}
-	}
-
-	return map[string]any{
-		"task_id": normalizedTaskID,
-		"type":    method,
-		"payload": payload,
-	}
-}
-
-// collectTaskIDs walks arbitrary decoded response payloads and gathers every
-// embedded task_id.
-func collectTaskIDs(rawValue any, ids map[string]struct{}) {
-	switch value := rawValue.(type) {
-	case map[string]any:
-		for key, item := range value {
-			if strings.HasSuffix(key, "task_id") {
-				if taskID, ok := item.(string); ok && taskID != "" {
-					ids[taskID] = struct{}{}
-				}
-			}
-			collectTaskIDs(item, ids)
-		}
-	case []map[string]any:
-		for _, item := range value {
-			collectTaskIDs(item, ids)
-		}
-	case []any:
-		for _, item := range value {
-			collectTaskIDs(item, ids)
-		}
-	}
 }
